Checkpointing

+++ controlled_sampling.py
@@ -147,7 +147,7 @@ def main():
         time = jax.random.uniform(keys[0], [1], minval=min_time, maxval=1.0)
         x = jax.random.normal(subkey, [1, *diffusion_model.shape])
         total_loss = 0.0
-        keys = jax.random.split(key, num=max_steps)
+        keys = jax.random.split(key, num=max_steps.astype(int))
         for i in jnp.arange(max_steps):
             x, time, clip_loss, control_loss = sample_step(policy_model, params, keys[i], x, time, target)
             total_loss += clip_loss + control_loss
